Use existing integration test
